% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/shared_methods.R
\name{compare_fn}
\alias{compare_fn}
\alias{compare_fn.r2r_hashtable}
\title{Get key comparison function of an hash table}
\usage{
compare_fn(x)

\method{compare_fn}{r2r_hashtable}(x)
}
\arguments{
\item{x}{an \code{hashset} or \code{hashmap}.}
}
\value{
a function.
}
\description{
Returns the key comparison function of an hash table
(\code{hashset} or \code{hashmap}).
}
\examples{
s <- hashset()
compare_fn(s)
}
\author{
Valerio Gherardi
}
