% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/hashmap.R
\name{default}
\alias{default}
\alias{default<-}
\alias{default.r2r_hashmap}
\alias{default<-.r2r_hashmap}
\title{Default \code{hashmap} values}
\usage{
default(x)

default(x) <- value

\method{default}{r2r_hashmap}(x)

\method{default}{r2r_hashmap}(x) <- value
}
\arguments{
\item{x}{an \code{hashmap}.}

\item{value}{an arbitrary R object. Default value to be associated to missing
keys in the \code{hashmap}.}
}
\value{
an arbitrary R objec.
}
\description{
These generics are used to get or set the default value of an
\code{hashmap}, optionally returned upon query of a missing key.
}
\details{
For more details, see the \link{hashtable} documentation page.
}
\examples{
m <- hashmap()
default(m)
default(m) <- 840
}
\author{
Valerio Gherardi
}
