% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{get_env_key}
\alias{get_env_key}
\title{Utility for collision handling}
\usage{
get_env_key(env, key, hash, compare)
}
\arguments{
\item{env}{an environment.}

\item{key}{an arbitrary R object.}
}
\value{
a character of length one. Either an hash \code{h} of the input key,
such that \code{identical(env[[h]], key)} is \code{TRUE}, or a name which is
unbind in \code{env}, if no match is found.
}
\description{
Utility for collision handling
}
