% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/shared_methods.R
\name{hash_fn}
\alias{hash_fn}
\alias{hash_fn.r2r_hashtable}
\title{Get hash function of an hash table}
\usage{
hash_fn(x)

\method{hash_fn}{r2r_hashtable}(x)
}
\arguments{
\item{x}{an \code{hashset} or \code{hashmap}.}
}
\value{
a function.
}
\description{
Returns the hash function used for key hashing in an hash table
(\code{hashset} or \code{hashmap}).
}
\examples{
s <- hashset()
hash_fn(s)
}
\author{
Valerio Gherardi
}
