% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R, R/generics.R, R/hashmap.R
\name{hashtable_methods}
\alias{hashtable_methods}
\alias{insert}
\alias{query}
\alias{delete}
\alias{has_key}
\alias{keys}
\alias{values}
\alias{\%has_key\%}
\alias{insert.r2r_hashmap}
\alias{delete.r2r_hashmap}
\alias{query.r2r_hashmap}
\alias{length.r2r_hashmap}
\alias{has_key.r2r_hashmap}
\alias{keys.r2r_hashmap}
\alias{values.r2r_hashmap}
\alias{[[.r2r_hashmap}
\alias{[.r2r_hashmap}
\alias{[[<-.r2r_hashmap}
\alias{[<-.r2r_hashmap}
\title{Methods for S3 classes \code{map} and \code{set}}
\usage{
insert(x, key, ...)

query(x, key)

delete(x, key)

has_key(x, key)

keys(x)

values(x)

x \%has_key\% key

\method{insert}{r2r_hashmap}(x, key, value, ...)

\method{delete}{r2r_hashmap}(x, key, ...)

\method{query}{r2r_hashmap}(x, key)

\method{length}{r2r_hashmap}(x)

\method{has_key}{r2r_hashmap}(x, key)

\method{keys}{r2r_hashmap}(x)

\method{values}{r2r_hashmap}(x)

\method{[[}{r2r_hashmap}(x, i)

\method{[}{r2r_hashmap}(x, i)

\method{[[}{r2r_hashmap}(x, i) <- value

\method{[}{r2r_hashmap}(x, i) <- value
}
\description{

}
\author{
Valerio Gherardi
}
