% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/hashmap.R, R/shared_methods.R
\name{hashtable_properties}
\alias{hashtable_properties}
\alias{hash_fn}
\alias{compare_fn}
\alias{on_missing_key}
\alias{on_missing_key<-}
\alias{default}
\alias{default<-}
\alias{on_missing_key.r2r_hashmap}
\alias{on_missing_key<-.r2r_hashmap}
\alias{default.r2r_hashmap}
\alias{default<-.r2r_hashmap}
\alias{hash_fn.r2r_hashtable}
\alias{compare_fn.r2r_hashtable}
\title{Hash table properties}
\usage{
hash_fn(x)

compare_fn(x)

on_missing_key(x)

on_missing_key(x, action) <- value

default(x)

default(x) <- value

\method{on_missing_key}{r2r_hashmap}(x)

\method{on_missing_key}{r2r_hashmap}(x, action) <- value

\method{default}{r2r_hashmap}(x)

\method{default}{r2r_hashmap}(x) <- value

\method{hash_fn}{r2r_hashtable}(x)

\method{compare_fn}{r2r_hashtable}(x)
}
\arguments{
\item{x}{an \code{hashset} or \code{hashmap}.}

\item{action}{a string, either \code{"throw"} or \code{"default"}. Action to
be taken upon query of a missing key.}

\item{value}{an arbitrary R object. Default value to be associated to missing
keys in the \code{hashmap}.}
}
\description{
This generics are used to get and (where allowed) set various
properties of \code{hashset} and \code{hashmap} objects.
}
\author{
Valerio Gherardi
}
