% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/shared_methods.R
\name{keys}
\alias{keys}
\alias{keys.r2r_hashtable}
\title{List all keys from an hash table}
\usage{
keys(x)

\method{keys}{r2r_hashtable}(x)
}
\arguments{
\item{x}{an \code{hashset} or \code{hashmap}.}
}
\value{
a list. Registered keys in the hash table \code{x}.
}
\description{
These generics are used for listing all keys registered in an
\code{hashset} or \code{hashmap}, respectively.
}
\author{
Valerio Gherardi
}
