% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_methods.R
\name{length.r2r_hashtable}
\alias{length.r2r_hashtable}
\title{Size of hash tables}
\usage{
\method{length}{r2r_hashtable}(x)
}
\arguments{
\item{x}{an \code{hashset} or \code{hashmap}.}
}
\value{
an integer. Number of keys in the hash table (or elements in a set).
}
\description{
Returns the total number of keys in an hash table.
}
\examples{
s <- hashset()
insert(s, "foo")
length(s)
}
\author{
Valerio Gherardi
}
