% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/hashmap.R
\name{on_missing_key}
\alias{on_missing_key}
\alias{on_missing_key<-}
\alias{on_missing_key.r2r_hashmap}
\alias{on_missing_key<-.r2r_hashmap}
\title{On missing key behaviour}
\usage{
on_missing_key(x)

on_missing_key(x) <- value

\method{on_missing_key}{r2r_hashmap}(x)

\method{on_missing_key}{r2r_hashmap}(x) <- value
}
\arguments{
\item{x}{an \code{hashmap}.}

\item{value}{a string, either \code{"throw"} or \code{"default"}. Action to
be taken upon query of a missing key.}
}
\value{
a string, either \code{"throw"} or \code{"default"}.
}
\description{
These generics are used to get or set the behaviour of an
\code{hashmap} upon query of a missing key (currently, only an
\code{hashmap} method is implemented).
}
\details{
For more details, see the \link{hashtable} documentation page.
}
\examples{
m <- hashmap()
on_missing_key(m)
on_missing_key(m) <- "throw"
}
\author{
Valerio Gherardi
}
