% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2r-package.R
\docType{package}
\name{r2r-package}
\alias{r2r}
\alias{r2r-package}
\title{r2r: R-Object to R-Object Hash Maps}
\description{
Implementation of hash tables (hash sets and hash maps) in R, 
	featuring arbitrary R objects as keys, 
	arbitrary hash and key-comparison functions, 
	and customizable behaviour upon queries of missing keys.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/r2r}
  \item Report bugs at \url{https://github.com/vgherard/r2r/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
