% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/hashmap.R
\name{values}
\alias{values}
\alias{values.r2r_hashmap}
\title{List all values from an hash map}
\usage{
values(x)

\method{values}{r2r_hashmap}(x)
}
\arguments{
\item{x}{an \code{hashset} or \code{hashmap}.}
}
\value{
a list. Values associated to keys in the hash map \code{x}.
}
\description{
This function is used to list all values associated to keys in
an \code{hashmap}. Implemented as a generic, but currently only the
\code{hashmap} method is defined.
}
\examples{
m <- hashmap(list("a", 1), list("b", 2))
values(m)
}
\author{
Valerio Gherardi
}
