% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{hashtable_methods}
\alias{hashtable_methods}
\title{Methods for S3 classes \code{hashmap} and \code{hashset}}
\description{
This page provides an overview of the available methods for
\code{hashmap} and \code{hashset} objects. We list methods based on the
general type of task addressed.
\subsection{Basic read/write operations}{
\itemize{
\item \code{\link[r2r]{insert}()}
\item \code{\link[r2r]{delete}()}
\item \code{\link[r2r]{query}()}
\item \link[r2r]{subsetting_hashtables}: \code{`[[`}, \code{`[[<-`}, \code{`[`}
and \code{`[<-`}
}
}

\subsection{Size of hash table}{
\itemize{
\item \code{length()}
}
}

\subsection{Other key or value access operations}{
\itemize{
\item \code{\link{keys}()}
\item \code{\link{values}()}
\item \code{\link{has_key}}, \code{\link[r2r:has_key]{\%has_key\%}}
}
}

\subsection{Get/set hashtable properties}{
\itemize{
\item \code{\link{hash_fn}()}
\item \code{\link{compare_fn}()}
\item \code{\link{on_missing_key}()}
\item \code{\link{default}()}
}
}
}
\author{
Valerio Gherardi
}
