% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{hashtable_properties}
\alias{hashtable_properties}
\title{Hash table properties}
\arguments{
\item{x}{an \code{hashset} or \code{hashmap}.}
}
\description{
This generics are used to get and (where allowed) set various
properties of \code{hashset} and \code{hashmap} objects.
}
\author{
Valerio Gherardi
}
